Tools: Migrated ZAxis and Tick3D to ES6 pattern.
