+++ js/masters/highcharts-3d.src.js
@@ -11,7 +11,9 @@
 import '../parts-3d/Math.js';
 import '../parts-3d/SVGRenderer.js';
 import '../parts-3d/Chart.js';
-import '../parts-3d/Axis.js';
+import '../parts-3d/ZAxis.js';
+import '../parts-3d/Axis3D.js';
+import '../parts-3d/Tick3D.js';
 import '../parts-3d/Series.js';
 import '../parts-3d/Column.js';
 import '../parts-3d/Pie.js';
