+++ js/parts-3d/Axis3D.js
@@ -15,117 +15,10 @@ import H from '../parts/Globals.js';
 import Tick from '../parts/Tick.js';
 import Tick3D from './Tick3D.js';
 import U from '../parts/Utilities.js';
-var addEvent = U.addEvent, extend = U.extend, merge = U.merge, pick = U.pick, splat = U.splat, wrap = U.wrap;
-import ZAxis from './ZAxis.js';
+var addEvent = U.addEvent, merge = U.merge, pick = U.pick, wrap = U.wrap;
 import '../parts/Chart.js';
-var Chart = H.Chart, deg2rad = H.deg2rad, perspective = H.perspective, perspective3D = H.perspective3D, shapeArea = H.shapeArea;
-/**
- * @optionparent xAxis
- */
-var extendedOptions = {
-    labels: {
-        /**
-         * Defines how the labels are be repositioned according to the 3D chart
-         * orientation.
-         *
-         * - `'offset'`: Maintain a fixed horizontal/vertical distance from the
-         *   tick marks, despite the chart orientation. This is the backwards
-         *   compatible behavior, and causes skewing of X and Z axes.
-         *
-         * - `'chart'`: Preserve 3D position relative to the chart.
-         *   This looks nice, but hard to read if the text isn't
-         *   forward-facing.
-         *
-         * - `'flap'`: Rotated text along the axis to compensate for the chart
-         *   orientation. This tries to maintain text as legible as possible
-         *   on all orientations.
-         *
-         * - `'ortho'`: Rotated text along the axis direction so that the labels
-         *   are orthogonal to the axis. This is very similar to `'flap'`,
-         *   but prevents skewing the labels (X and Y scaling are still
-         *   present).
-         *
-         * @sample highcharts/3d/skewed-labels/
-         *         Skewed labels
-         *
-         * @since      5.0.15
-         * @validvalue ['offset', 'chart', 'flap', 'ortho']
-         * @product    highcharts
-         * @requires   highcharts-3d
-         */
-        position3d: 'offset',
-        /**
-         * If enabled, the axis labels will skewed to follow the perspective.
-         *
-         * This will fix overlapping labels and titles, but texts become less
-         * legible due to the distortion.
-         *
-         * The final appearance depends heavily on `labels.position3d`.
-         *
-         * @sample highcharts/3d/skewed-labels/
-         *         Skewed labels
-         *
-         * @since    5.0.15
-         * @product  highcharts
-         * @requires highcharts-3d
-         */
-        skew3d: false
-    },
-    title: {
-        /**
-         * Defines how the title is repositioned according to the 3D chart
-         * orientation.
-         *
-         * - `'offset'`: Maintain a fixed horizontal/vertical distance from the
-         *   tick marks, despite the chart orientation. This is the backwards
-         *   compatible behavior, and causes skewing of X and Z axes.
-         *
-         * - `'chart'`: Preserve 3D position relative to the chart.
-         *   This looks nice, but hard to read if the text isn't
-         *   forward-facing.
-         *
-         * - `'flap'`: Rotated text along the axis to compensate for the chart
-         *   orientation. This tries to maintain text as legible as possible on
-         *   all orientations.
-         *
-         * - `'ortho'`: Rotated text along the axis direction so that the labels
-         *   are orthogonal to the axis. This is very similar to `'flap'`, but
-         *   prevents skewing the labels (X and Y scaling are still present).
-         *
-         * - `undefined`: Will use the config from `labels.position3d`
-         *
-         * @sample highcharts/3d/skewed-labels/
-         *         Skewed labels
-         *
-         * @type     {"offset"|"chart"|"flap"|"ortho"|null}
-         * @since    5.0.15
-         * @product  highcharts
-         * @requires highcharts-3d
-         */
-        position3d: null,
-        /**
-         * If enabled, the axis title will skewed to follow the perspective.
-         *
-         * This will fix overlapping labels and titles, but texts become less
-         * legible due to the distortion.
-         *
-         * The final appearance depends heavily on `title.position3d`.
-         *
-         * A `null` value will use the config from `labels.skew3d`.
-         *
-         * @sample highcharts/3d/skewed-labels/
-         *         Skewed labels
-         *
-         * @type     {boolean|null}
-         * @since    5.0.15
-         * @product  highcharts
-         * @requires highcharts-3d
-         */
-        skew3d: null
-    }
-};
+var deg2rad = H.deg2rad, perspective = H.perspective, perspective3D = H.perspective3D, shapeArea = H.shapeArea;
 /* eslint-disable no-invalid-this */
-merge(true, Axis.defaultOptions, extendedOptions);
 addEvent(Axis, 'afterSetOptions', function () {
     var options;
     if (this.chart.is3d && this.chart.is3d() && this.coll !== 'colorAxis') {
@@ -238,29 +131,6 @@ addEvent(Axis, 'destroy', function () {
             this[prop] = this[prop].destroy();
         }
     }, this);
-});
-/*
-Z-AXIS
- */
-Chart.prototype.addZAxis = function (options) {
-    return new ZAxis(this, options);
-};
-Chart.prototype.collectionsWithUpdate.push('zAxis');
-Chart.prototype.collectionsWithInit.zAxis = [Chart.prototype.addZAxis];
-// Get the Z axis in addition to the default X and Y.
-addEvent(Chart, 'afterGetAxes', function () {
-    var chart = this, options = this.options, zAxisOptions = options.zAxis = splat(options.zAxis || {});
-    if (!chart.is3d()) {
-        return;
-    }
-    this.zAxis = [];
-    zAxisOptions.forEach(function (axisOptions, i) {
-        axisOptions.index = i;
-        // Z-Axis is shown horizontally, so it's kind of a X-Axis
-        axisOptions.isX = true;
-        var zAxis = chart.addZAxis(axisOptions);
-        zAxis.setScale();
-    });
 });
 /* eslint-disable valid-jsdoc */
 /**
@@ -502,14 +372,15 @@ var Axis3D = /** @class */ (function () {
      * Extends axis class with 3D support.
      * @private
      */
-    Axis3D.compose = function (AxisClass, _ChartClass, TickClass) {
+    Axis3D.compose = function (AxisClass) {
+        merge(true, Axis.defaultOptions, Axis3D.defaultOptions);
         AxisClass.keepProps.push('axis3D');
         addEvent(AxisClass, 'init', Axis3D.onInit);
         var axisProto = AxisClass.prototype;
         // Wrap getSlotWidth function to calculate individual width value for
         // each slot (#8042).
         wrap(axisProto, 'getSlotWidth', Axis3D.wrapGetSlotWidth);
-        Tick3D.compose(TickClass);
+        Tick3D.compose(Tick);
     };
     /**
      * @private
@@ -577,7 +448,118 @@ var Axis3D = /** @class */ (function () {
         }
         return proceed.apply(axis, [].slice.call(arguments, 1));
     };
+    /* *
+     *
+     *  Static Properties
+     *
+     * */
+    /**
+     * @optionparent xAxis
+     */
+    Axis3D.defaultOptions = {
+        labels: {
+            /**
+             * Defines how the labels are be repositioned according to the 3D
+             * chart orientation.
+             *
+             * - `'offset'`: Maintain a fixed horizontal/vertical distance from
+             *   the tick marks, despite the chart orientation. This is the
+             *   backwards compatible behavior, and causes skewing of X and Z
+             *   axes.
+             *
+             * - `'chart'`: Preserve 3D position relative to the chart. This
+             *   looks nice, but hard to read if the text isn't forward-facing.
+             *
+             * - `'flap'`: Rotated text along the axis to compensate for the
+             *   chart orientation. This tries to maintain text as legible as
+             *   possible on all orientations.
+             *
+             * - `'ortho'`: Rotated text along the axis direction so that the
+             *   labels are orthogonal to the axis. This is very similar to
+             *   `'flap'`, but prevents skewing the labels (X and Y scaling are
+             *   still present).
+             *
+             * @sample highcharts/3d/skewed-labels/
+             *         Skewed labels
+             *
+             * @since      5.0.15
+             * @validvalue ['offset', 'chart', 'flap', 'ortho']
+             * @product    highcharts
+             * @requires   highcharts-3d
+             */
+            position3d: 'offset',
+            /**
+             * If enabled, the axis labels will skewed to follow the
+             * perspective.
+             *
+             * This will fix overlapping labels and titles, but texts become
+             * less legible due to the distortion.
+             *
+             * The final appearance depends heavily on `labels.position3d`.
+             *
+             * @sample highcharts/3d/skewed-labels/
+             *         Skewed labels
+             *
+             * @since    5.0.15
+             * @product  highcharts
+             * @requires highcharts-3d
+             */
+            skew3d: false
+        },
+        title: {
+            /**
+             * Defines how the title is repositioned according to the 3D chart
+             * orientation.
+             *
+             * - `'offset'`: Maintain a fixed horizontal/vertical distance from
+             *   the tick marks, despite the chart orientation. This is the
+             *   backwards compatible behavior, and causes skewing of X and Z
+             *   axes.
+             *
+             * - `'chart'`: Preserve 3D position relative to the chart. This
+             *   looks nice, but hard to read if the text isn't forward-facing.
+             *
+             * - `'flap'`: Rotated text along the axis to compensate for the
+             *   chart orientation. This tries to maintain text as legible as
+             *   possible on all orientations.
+             *
+             * - `'ortho'`: Rotated text along the axis direction so that the
+             *   labels are orthogonal to the axis. This is very similar to
+             *   `'flap'`, but prevents skewing the labels (X and Y scaling are
+             *   still present).
+             *
+             * - `undefined`: Will use the config from `labels.position3d`
+             *
+             * @sample highcharts/3d/skewed-labels/
+             *         Skewed labels
+             *
+             * @type     {"offset"|"chart"|"flap"|"ortho"|null}
+             * @since    5.0.15
+             * @product  highcharts
+             * @requires highcharts-3d
+             */
+            position3d: null,
+            /**
+             * If enabled, the axis title will skewed to follow the perspective.
+             *
+             * This will fix overlapping labels and titles, but texts become
+             * less legible due to the distortion.
+             *
+             * The final appearance depends heavily on `title.position3d`.
+             *
+             * A `null` value will use the config from `labels.skew3d`.
+             *
+             * @sample highcharts/3d/skewed-labels/
+             *         Skewed labels
+             *
+             * @type     {boolean|null}
+             * @since    5.0.15
+             * @product  highcharts
+             * @requires highcharts-3d
+             */
+            skew3d: null
+        }
+    };
     return Axis3D;
 }());
-Axis3D.compose(Axis, Chart, Tick);
 export default Axis3D;
