+++ ts/parts/Axis.ts
@@ -11,6 +11,7 @@
 'use strict';
 
 import type { AxisComposition } from './axis/types';
+import type ZAxis from '../parts-3d/ZAxis';
 import Color from './Color.js';
 import H from './Globals.js';
 import Tick from './Tick.js';
@@ -56,7 +57,7 @@ declare global {
             'scrollbar'|'traverseUpButton'|'zoom'
         );
         type AxisMinorTickPositionValue = ('inside'|'outside');
-        type AxisOptions = (XAxisOptions|YAxisOptions|ZAxisOptions);
+        type AxisOptions = (XAxisOptions|YAxisOptions|ZAxis.Options);
         type AxisTickmarkPlacementValue = ('between'|'on');
         type AxisTickPositionValue = ('inside'|'outside');
         type AxisTitleAlignValue = ('high'|'low'|'middle');
@@ -297,13 +298,6 @@ declare global {
             staticScale?: number;
             stops?: Array<GradientColorStopObject>;
             tooltipValueFormat?: string;
-        }
-        interface ZAxisOptions extends XAxisOptions {
-            breaks?: undefined;
-            crosshair?: undefined;
-            lineColor?: undefined;
-            lineWidth?: undefined;
-            showEmpty?: undefined;
         }
         class Axis {
             public static defaultBottomAxisOptions: AxisOptions;
