+++ ts/parts/axis/types.d.ts
@@ -9,6 +9,7 @@
  * */
 
 import type Axis from '../Axis';
+import type Axis3D from '../../parts-3d/Axis3D';
 import type BrokenAxis from '../../modules/broken-axis.src';
 import type DateTimeAxis from '../DateTimeAxis';
 import type GridAxis from '../../parts-gantt/GridAxis';
@@ -20,6 +21,7 @@ import type OrdinalAxis from '../OrdinalAxis';
 import type ParallelAxis from '../../modules/parallel-coordinates.src';
 import type RadialAxis from '../../parts-more/RadialAxis';
 import type ScrollbarAxis from '../ScrollbarAxis';
+import type ZAxis from '../../parts-3d/ZAxis';
 
 export interface AxisBreakBorderObject {
     move: string;
@@ -34,6 +36,7 @@ export interface AxisBreakObject {
 }
 
 export interface AxisComposition {
+    axis3D?: Axis3D['axis3D'];
     brokenAxis?: BrokenAxis['brokenAxis'];
     dateTime?: DateTimeAxis['dateTime'];
     grid?: GridAxis['grid'];
@@ -50,6 +53,7 @@ export interface AxisComposition {
  */
 export type AxisType = (
     Axis|
+    Axis3D|
     BrokenAxis|
     GridAxis|
     HiddenAxis|
@@ -59,5 +63,6 @@ export type AxisType = (
     OrdinalAxis|
     ParallelAxis|
     RadialAxis|
-    ScrollbarAxis
+    ScrollbarAxis|
+    ZAxis
 );
